requirements.txt
aniso8601==8.1.1
arabic-reshaper==2.1.3
asgiref==3.3.4
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
Faker==8.4.0
firebase-admin==5.0.0
future==0.18.2
google-api-core==1.28.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.5
graphql-relay==3.1.0
grpcio==1.38.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
jsonfield==3.1.0
msgpack==1.0.2
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
promise==2.3
proto-plus==1.18.1
protobuf==3.17.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
